Experiment with Oklab color space in DEBUG builds.
Oklab color space is meant to maintain luminance as hue shifts, so on
the surface it sounds like it would be great for a color wheel.

But Oklab has a compressed yellow range and a compressed cyan range.
And yellow and cyan only really show up at high luminance.

It may be great for maintaining hue as luminance changes, but it's not
great for getting adequate representation of hues in a color wheel with
medium luminance.

So, it's only present for experimentation, not for actual use.